Flickering hotfix
I think it's usable now. It's a bit slower, than serial, because the TCP protocol-stack.

Now on ESP32 and ESP8266 it uses a buffer when the redrawLine function is used. This way the whole packet goes out to the client in one peace. It solved the flickering with me.

Commander compatibility needs to be tested.