In-house websocket server implementation prototype
From now on, the in-house version of the websocket server would be the preferred way to go. It remowes library dependencies and it can be optimized to text only operation( whisch is all we need ). It is a reduced implementation and only works with 125 byte long payloads, but I think it would be sufficient four our requirements. It is not finished, the bindings for the Stream interface needs to be done. Also the base Shellminator class needs to be modified to support this implementation.