WebServer class created with example. Conda environment
It wass neccesary, because Arduino environment does not has a multi-platform webserver implementation. The current implementation is not the fastest, because it is not aync, but it runs on Uno R4, ESP32, ESP8266 and Pi Pico( probably more as well, but these are the tested ones ).

I worked on the hosted webpage design as well, I think it is a bit more modern now.